tweak speed
the old system could result in negative frequencies

in practice, the speed is limited by the keyboard repeat rate. But we
don't know that. A typical value is 30/s

+++ items.go
@@ -16,7 +16,7 @@ type Item struct {
 	Attack      float64
 	Defense     float64
 	LineOfSight int
-	Speed       float64
+	Speed       int
 }
 
 var Items = map[string]Item{
